refactor(dashboard): promote shared lens helpers to lenscommon
Move multi-chart helpers (Lens filters, chart filter JSON, FilterSimple,
JSON-with-defaults marshal/preserve, canonical JSON key sorting) into
lenscommon with exported APIs; keep dashboard wrappers for legacy call sites.
