+++ internal/kibana/dashboard/lenscommon/filter_simple.go
@@ -15,18 +15,37 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package dashboard
+package lenscommon
 
 import (
 	"github.com/elastic/terraform-provider-elasticstack/generated/kbapi"
-	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/lenscommon"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/models"
+	"github.com/elastic/terraform-provider-elasticstack/internal/utils/typeutils"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
-func filterSimpleFromAPI(m *models.FilterSimpleModel, apiQuery kbapi.FilterSimple) {
-	lenscommon.FilterSimpleFromAPI(m, apiQuery)
+// FilterSimpleFromAPI maps kbapi.FilterSimple into FilterSimpleModel.
+func FilterSimpleFromAPI(m *models.FilterSimpleModel, apiQuery kbapi.FilterSimple) {
+	m.Expression = types.StringValue(apiQuery.Expression)
+	if apiQuery.Language == nil {
+		m.Language = types.StringValue(string(kbapi.FilterSimpleLanguageKql))
+		return
+	}
+	m.Language = typeutils.StringishPointerValue(apiQuery.Language)
 }
 
-func filterSimpleToAPI(m *models.FilterSimpleModel) kbapi.FilterSimple {
-	return lenscommon.FilterSimpleToAPI(m)
+// FilterSimpleToAPI maps FilterSimpleModel into kbapi.FilterSimple.
+func FilterSimpleToAPI(m *models.FilterSimpleModel) kbapi.FilterSimple {
+	if m == nil {
+		return kbapi.FilterSimple{}
+	}
+
+	query := kbapi.FilterSimple{
+		Expression: m.Expression.ValueString(),
+	}
+	if typeutils.IsKnown(m.Language) {
+		lang := kbapi.FilterSimpleLanguage(m.Language.ValueString())
+		query.Language = &lang
+	}
+	return query
 }
