+++ internal/kibana/dashboard/models_lens_chart_helpers.go
@@ -24,7 +24,6 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/generated/kbapi"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/lenscommon"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/models"
-	"github.com/elastic/terraform-provider-elasticstack/internal/kibana/dashboard/panelkit"
 	"github.com/elastic/terraform-provider-elasticstack/internal/utils/customtypes"
 	"github.com/hashicorp/terraform-plugin-framework-jsontypes/jsontypes"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
@@ -36,38 +35,13 @@ import (
 // populateFiltersFromAPI converts a slice of kbapi.LensPanelFilters_Item into models.ChartFilterJSONModel
 // values, appending any errors to diags.
 func populateFiltersFromAPI(filters []kbapi.LensPanelFilters_Item, diags *diag.Diagnostics) []models.ChartFilterJSONModel {
-	if len(filters) == 0 {
-		return nil
-	}
-	result := make([]models.ChartFilterJSONModel, 0, len(filters))
-	for _, f := range filters {
-		fm := models.ChartFilterJSONModel{}
-		fd := chartFilterJSONPopulateFromAPIItem(&fm, f)
-		diags.Append(fd...)
-		if !fd.HasError() {
-			result = append(result, fm)
-		}
-	}
-	return result
+	return lenscommon.PopulateFiltersFromAPI(filters, diags)
 }
 
 // buildFiltersForAPI converts the model filter slice into the kbapi type, appending errors to diags.
 // The returned slice is always non-nil (empty API payload is []kbapi.LensPanelFilters_Item{}).
 func buildFiltersForAPI(filters []models.ChartFilterJSONModel, diags *diag.Diagnostics) []kbapi.LensPanelFilters_Item {
-	if len(filters) == 0 {
-		return []kbapi.LensPanelFilters_Item{}
-	}
-
-	items := make([]kbapi.LensPanelFilters_Item, 0, len(filters))
-	for _, f := range filters {
-		var item kbapi.LensPanelFilters_Item
-		fd := decodeChartFilterJSON(f.FilterJSON, &item)
-		diags.Append(fd...)
-		if !fd.HasError() {
-			items = append(items, item)
-		}
-	}
-	return items
+	return lenscommon.BuildFiltersForAPI(filters, diags)
 }
 
 // marshalToNormalized delegates to lenscommon.MarshalToNormalized (canonical implementation).
@@ -83,15 +57,11 @@ func preservePriorNormalizedWithDefaultsIfEquivalent[T any](ctx context.Context,
 // marshalToJSONWithDefaults stores the already-marshaled bytes as a JSONWithDefaultsValue,
 // or adds an error to diags and returns (zero, false) on failure.
 func marshalToJSONWithDefaults[T any](bytes []byte, err error, fieldName string, defaults func(T) T, diags *diag.Diagnostics) (customtypes.JSONWithDefaultsValue[T], bool) {
-	if err != nil {
-		diags.AddError("Failed to marshal "+fieldName, err.Error())
-		return customtypes.JSONWithDefaultsValue[T]{}, false
-	}
-	return customtypes.NewJSONWithDefaultsValue(string(bytes), defaults), true
+	return lenscommon.MarshalToJSONWithDefaults(bytes, err, fieldName, defaults, diags)
 }
 
 func preservePriorJSONWithDefaultsIfEquivalent[T any](ctx context.Context, prior, current customtypes.JSONWithDefaultsValue[T], diags *diag.Diagnostics) customtypes.JSONWithDefaultsValue[T] {
-	return panelkit.PreservePriorJSONWithDefaultsIfEquivalent(ctx, prior, current, diags)
+	return lenscommon.PreservePriorJSONWithDefaultsIfEquivalent(ctx, prior, current, diags)
 }
 
 // lensDataSourceIsESQLOrTable reports whether a Lens chart's `data_source` union
